% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheMemoiser.R
\name{diskCache}
\alias{diskCache}
\title{Disk based Memo-Cache}
\usage{
diskCache(filename = NULL)
}
\arguments{
\item{filename}{name of the file to load the cache from, or store it into.}
}
\value{
cache memoiser object
}
\description{
Creates a new disk based cache object.
}
\details{
Originally from Hadley Wickham's memoise package.
  changes: added load and save to disk
}

