% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_memo.R
\name{redis.memoiser.flush}
\alias{redis.memoiser.flush}
\title{Delete all memoized information in redis DB}
\usage{
redis.memoiser.flush(host = "127.0.0.1", port = 6379)
}
\arguments{
\item{host}{Redis host IP}

\item{port}{Redis host port number}
}
\description{
Delete all memoized information in redis DB
}

