% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_memo.R
\name{redis.threadsafe.memoiser}
\alias{redis.threadsafe.memoiser}
\title{Memoisaztion using redis for parallel environment}
\usage{
redis.threadsafe.memoiser(fun, use.func.contents = FALSE,
  compare.args.as.characters = FALSE, redis.host = "127.0.0.1",
  redis.port = 6379)
}
\arguments{
\item{fun}{The funciton to memoise}

\item{use.func.contents}{Boolean, to use the function contents (using deparse) or simply use its character name to store the results in DB}

\item{compare.args.as.characters}{Boolean, to compare the function input arguements in binary or as character}

\item{redis.host}{Redis host IP}

\item{redis.port}{Redis host port number}
}
\value{
A memoised function. When called, connect to redis to see if results are available. If not, will run the function to get results.
}
\description{
In parallel environment, one needs to connect to redis again. This function does this automatically and returns a memoised function.
}

