% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diskMemoiser.R
\name{diskMemoiser}
\alias{diskMemoiser}
\alias{diskMemoizer}
\title{Disk Memoisation}
\usage{
diskMemoiser(fun, memo.dir = "./memo", use.func.contents = FALSE,
  compare.args.as.characters = FALSE)
}
\arguments{
\item{fun}{An arbitrary function}

\item{memo.dir}{An optional directory to store the memoised results. 
If not specified, defaults to "./memo"}

\item{use.func.contents}{If TRUE, will hash \code{fun}'s contents and
use it for looking up for the memoized object.
Otherwise, function name will be used.
Useful when functions are anonymous or their contents may
change.}

\item{compare.args.as.characters}{If TRUE, converts args to a character
string before comparing them via memoized objects.
Useful when args may vary in their current form 
(e.g., pointers)}
}
\value{
A memoised function
}
\description{
Performs memoisation of arbitrary functions to disk to speed up repeated 
calls to those functions.
}
\details{
This function defines a memoisation wrapper which saves the results of the 
function to disk. If the function is called with the same parameters it 
retrieves the previous result and returns it. Note that this only makes sense
if the function is a deterministic and not dependent on external variables. 
Use carefully!
}
\examples{
myfunction.slow <- function(x, y) {
  Sys.sleep(1)
  return(x + sin(y))
}

myfunction <- diskMemoiser(myfunction.slow)
system.time({ result.slow <- myfunction(3, 5) })
system.time({ result.fast <- myfunction(3, 5) })

# The memoization directory defaults to ./memo, but can be reassigned:
myfunction <- diskMemoiser(myfunction.slow, memo.dir = "./temp")
result.slow <- myfunction(7, 11)
result.fast <- myfunction(7, 11)
}
\references{
Based on \url{http://cran.r-project.org/web/packages/memoise}.
}

